requirements.txt

python-dotenv>=1.0.1
streamlit>=1.36.0
llama-index-core>=0.11.0
llama-index-vector-stores-pinecone>=0.3.0
llama-index-embeddings-nvidia>=0.1.0
llama-index-llms-nvidia>=0.1.0
pinecone-client>=3.2.2
pymongo>=4.6.0
dnspython>=2.6.1
pydantic>=2.5.0
openai>=1.40.0
# Optional readers for ingestion formats
# llama-index-readers-file>=0.2.0
# pypdf
# python-docx
# unstructured
# pillow
